Fanspin!
As famously known: We have fanspin!

A number of things were done, including:
 - added some logging
 - fixed session building (nasty bug):
    The problem ended up being a bug in the persistence
    logic identity key ID would be stored as 0 (which is
    wrong) and the Signal library in rust would try to
    check if the identity key is trusted by calling into
    kotlin code via JNI. It was returning null for the
    identity key and this is why get_string on obj returned
    a null pointer. Sadly I had to find this out by using
    log messages as neither the stack trace nor the error
    message provided any information about what
    exactly is null

Message encryption/decryption partially works now.
There are still bugs to squash and things to implement.
But at least something works.

+++ app/src/main/java/dev/superboring/aosp/chakonati/persistence/entities/DBMessage.kt
@@ -18,7 +18,7 @@ import dev.superboring.aosp.chakonati.components.fragments.chat.MessageFrom
 )
 data class DBMessage(
     @ColumnInfo(name = "id") @PrimaryKey(autoGenerate = true) val id: Long = 0,
-    @ColumnInfo(name = "chat_id") val chatId: Int = 0,
+    @ColumnInfo(name = "chat_id") val chatId: Long = 0,
     @ColumnInfo(name = "from") val messageFrom: Int,
     @ColumnInfo(name = "sender") val sender: String? = null,
     @ColumnInfo(name = "message_text") val messageText: String,
